argc and argv commits

+++ 0x0A-argc_argv/3-mul.c
@@ -13,11 +13,11 @@ int main(int argc, char **argv)
 if (argc > 1)
 {
 int i;
-int mul;
+long mul;
 mul = 1;
 for (i = 1; i < argc; i++)
 mul *= atoi(argv[i]);
-printf("%d\n", mul);
+printf("%ld\n", mul);
 }
 else
 {
